@@ -28,8 +28,8 @@
     "normalize.css": "^8.0.1",
     "prismjs": "^1.15.0",
     "prop-types": "^15.6.2",
-    "react": "^16.5.1",
-    "react-dom": "^16.5.1",
+    "react": "^16.8.3",
+    "react-dom": "^16.8.3",
     "react-helmet": "^5.2.0"
   },
   "repository": {
