@@ -1,31 +1,45 @@
 /* eslint-disable no-script-url */
-import React from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { Link as InternalLink } from 'gatsby';
 import FancyLink from './FancyLink';
 
 import CSS from '../css/PostHeader.module.css';
 
-class PostHeader extends React.Component {
-  stickyRef = React.createRef();
+function PostHeader({ post, listData }) {
+  const [{ prevPost, nextPost }, setPosts] = useState({ prevPost: false, nextPost: false });
 
-  fakeStickyRef = React.createRef();
+  const stickyRef = useRef();
+  const fakeStickyRef = useRef();
 
-  state = {
-    prevPost: false,
-    nextPost: false,
+  const handleScroll = () => {
+    if (stickyRef.current.classList.contains(CSS.stickyHeader)) {
+      // If it contains it, we're looking to remove it.
+      // Remove it if our scroll hits the top of where it should be.
+      if (window.pageYOffset < fakeStickyRef.current.offsetTop) {
+        stickyRef.current.classList.remove(CSS.stickyHeader);
+        fakeStickyRef.current.style.display = 'none';
+      } // Else do nothing.
+    } else if (window.pageYOffset > stickyRef.current.offsetTop) {
+      // If it does not contain it, we're looking to add it.
+      // Add it if our scroll hits the top of where it is.
+      stickyRef.current.classList.add(CSS.stickyHeader);
+      fakeStickyRef.current.style.display = 'block';
+    }
   };
 
-  componentDidMount = () => {
-    // On mount, calculate the prev/next links.
-    const { post: currPost } = this.props;
-    const { listData } = this.props;
+  const scrollToTop = () => {
+    document.body.scrollTop = 0;
+    document.documentElement.scrollTop = 0;
+  };
 
+  useEffect(() => {
+    // On mount, calculate the prev/next links.
     // Loop through the listData, when we find our post we can infer prev and next links from the list.
     for (let i = 0; i < listData.length; ++i) {
-      if (listData[i].node.id === currPost.id) {
+      if (listData[i].node.id === post.id) {
         // When we find ours, add the previous and next then break (we only need one of each per page).
-        this.setState({
+        setPosts({
           prevPost: (listData[i - 1] && listData[i - 1].node) || false,
           nextPost: (listData[i + 1] && listData[i + 1].node) || false,
         });
@@ -35,92 +49,63 @@ class PostHeader extends React.Component {
     }
 
     // And add scroll handler for header.
-    window.addEventListener('scroll', this.handleScroll, false);
-  };
+    window.addEventListener('scroll', handleScroll, false);
 
-  componentWillUnmount = () => {
-    window.removeEventListener('scroll', this.handleScroll, false);
-  };
+    return (() => window.removeEventListener('scroll', handleScroll, false));
+  }, []);
 
-  handleScroll = () => {
-    const { current: sticky } = this.stickyRef;
-    const { current: fakeSticky } = this.fakeStickyRef;
-    if (sticky.classList.contains(CSS.stickyHeader)) {
-      // If it contains it, we're looking to remove it.
-      // Remove it if our scroll hits the top of where it should be.
-      if (window.pageYOffset < fakeSticky.offsetTop) {
-        sticky.classList.remove(CSS.stickyHeader);
-        fakeSticky.style.display = 'none';
-      } // Else do nothing.
-    } else if (window.pageYOffset > sticky.offsetTop) {
-      // If it does not contain it, we're looking to add it.
-      // Add it if our scroll hits the top of where it is.
-      sticky.classList.add(CSS.stickyHeader);
-      fakeSticky.style.display = 'block';
-    }
-  };
-
-  scrollToTop = () => {
-    document.body.scrollTop = 0;
-    document.documentElement.scrollTop = 0;
-  };
-
-  render = () => {
-    const { post } = this.props;
-    const { prevPost, nextPost } = this.state;
-    return (
-      <header className={CSS.postHeader}>
-        <div className={CSS.topHeader}>
-          <InternalLink
-            to={(prevPost && prevPost.frontmatter.path) || 'javascript:;'}
-            className={`${CSS.leftArrowLink} ${!prevPost && CSS.disabled}`}
-            title="Last Post"
-          >
-            <div className={CSS.arrow}>
-              <div className={CSS.leftArrow} />
-            </div>
-          </InternalLink>
-          <div className={CSS.nameContainer}>
-            <FancyLink to="/" internal animated>
-              Noah Yamamoto
-            </FancyLink>
-          </div>
-          <InternalLink
-            to={(nextPost && nextPost.frontmatter.path) || 'javascript:;'}
-            className={`${CSS.rightArrowLink} ${!nextPost && CSS.disabled}`}
-            title="Next Post"
-          >
-            <div className={CSS.arrow}>
-              <div className={CSS.rightArrow} />
-            </div>
-          </InternalLink>
-        </div>
-        <div
-          className={CSS.bottomHeader}
-          style={{ display: 'none', visibility: 'hidden' }}
-          ref={this.fakeStickyRef}
+  return (
+    <header className={CSS.postHeader}>
+      <div className={CSS.topHeader}>
+        <InternalLink
+          to={(prevPost && prevPost.frontmatter.path) || 'javascript:;'}
+          className={`${CSS.leftArrowLink} ${!prevPost && CSS.disabled}`}
+          title="Last Post"
         >
-          <h2 className={CSS.postTitle}>{post.frontmatter.title}</h2>
-          <p className={CSS.postDate}>{post.frontmatter.date}</p>
+          <div className={CSS.arrow}>
+            <div className={CSS.leftArrow} />
+          </div>
+        </InternalLink>
+        <div className={CSS.nameContainer}>
+          <FancyLink to="/" internal animated>
+            Noah Yamamoto
+          </FancyLink>
         </div>
-        <div className={CSS.bottomHeader} ref={this.stickyRef}>
-          <h2 className={CSS.postTitle}>
-            {/* eslint-disable-next-line */}
-            <span
-              onClick={this.scrollToTop}
-              alt="Scroll back to top"
-              role="button"
-              tabIndex={0}
-              title="Scroll back to top"
-            >
-              {post.frontmatter.title}
-            </span>
-          </h2>
-          <p className={CSS.postDate}>{post.frontmatter.date}</p>
-        </div>
-      </header>
-    );
-  };
+        <InternalLink
+          to={(nextPost && nextPost.frontmatter.path) || 'javascript:;'}
+          className={`${CSS.rightArrowLink} ${!nextPost && CSS.disabled}`}
+          title="Next Post"
+        >
+          <div className={CSS.arrow}>
+            <div className={CSS.rightArrow} />
+          </div>
+        </InternalLink>
+      </div>
+      <div
+        className={CSS.bottomHeader}
+        style={{ display: 'none', visibility: 'hidden' }}
+        ref={fakeStickyRef}
+      >
+        <h2 className={CSS.postTitle}>{post.frontmatter.title}</h2>
+        <p className={CSS.postDate}>{post.frontmatter.date}</p>
+      </div>
+      <div className={CSS.bottomHeader} ref={stickyRef}>
+        <h2 className={CSS.postTitle}>
+          {/* eslint-disable-next-line */}
+          <span
+            onClick={scrollToTop}
+            alt="Scroll back to top"
+            role="button"
+            tabIndex={0}
+            title="Scroll back to top"
+          >
+            {post.frontmatter.title}
+          </span>
+        </h2>
+        <p className={CSS.postDate}>{post.frontmatter.date}</p>
+      </div>
+    </header>
+  );
 }
 
 PostHeader.propTypes = {
