@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { graphql } from 'gatsby';
 import HeadTag from '../components/HeadTag';
@@ -33,27 +33,15 @@ export const pageQuery = graphql`
   }
 `;
 
+function projects({ data: { allProjectsJson: { edges } } }) {
+  const [imgLoaded, setLoaded] = useState(false);
 
-class projects extends React.Component {
-  stickyRef = React.createRef();
+  const stickyRef = useRef();
+  const fakeStickyRef = useRef();
 
-  fakeStickyRef = React.createRef();
-
-  state = {
-    imgLoaded: false,
-  };
-
-  componentDidMount = () => {
-    window.addEventListener('scroll', this.handleScroll, false);
-  }
-
-  componentWillUnmount = () => {
-    window.removeEventListener('scroll', this.handleScroll, false);
-  };
-
-  handleScroll = () => {
-    const { current: sticky } = this.stickyRef;
-    const { current: fakeSticky } = this.fakeStickyRef;
+  const handleScroll = () => {
+    const { current: sticky } = stickyRef;
+    const { current: fakeSticky } = fakeStickyRef;
     if (sticky.classList.contains(CSS.stickyHeader)) {
       // If it contains it, we're looking to remove it.
       // Remove it if our scroll hits the top of where it should be.
@@ -69,60 +57,61 @@ class projects extends React.Component {
     }
   };
 
-  render() {
-    const { data: { allProjectsJson: { edges } } } = this.props;
-    const { imgLoaded } = this.state;
-    return (
-      <section className={CSS.blog}>
-        <HeadTag title="Projects" description="Projects index for NoahYamamoto.com" path="/projects" />
-        <Canvas />
-        <header className={CSS.projectsHeader}>
-          <div className={CSS.topHeader}>
-            <div className={CSS.nameContainer}>
-              <FancyLink to="/" internal animated>
-                Noah Yamamoto
-              </FancyLink>
-            </div>
-          </div>
-          <div
-            className={CSS.bottomHeader}
-            style={{ display: 'none', visibility: 'hidden' }}
-            ref={this.fakeStickyRef}
-          >
-            <h2 className={CSS.postTitle}>Projects</h2>
-          </div>
-          <div className={CSS.bottomHeader} ref={this.stickyRef}>
-            <h2 className={CSS.postTitle}>Projects</h2>
-          </div>
-        </header>
-        <div className={CSS.content}>
-          <p className={CSS.indexBlurb}>
-            Below are a few personal projects I&apos;ve done that I&apos;m proud of, to see more check out my&nbsp;
-            <FancyLink to="https://github.com/egrodo" newTab>
-              Github page
+  useEffect(() => {
+    window.addEventListener('scroll', handleScroll, false);
+    return (() => window.removeEventListener('scroll', handleScroll, false));
+  }, []);
+
+  return (
+    <section className={CSS.blog}>
+      <HeadTag title="Projects" description="Projects index for NoahYamamoto.com" path="/projects" />
+      <Canvas />
+      <header className={CSS.projectsHeader}>
+        <div className={CSS.topHeader}>
+          <div className={CSS.nameContainer}>
+            <FancyLink to="/" internal animated>
+              Noah Yamamoto
             </FancyLink>
-            !
-          </p>
-          <div className={CSS.GHChart}>
-            <div className={`${CSS.imgPlaceholder} ${imgLoaded ? CSS.hide : ''}`}>
-              <Spinner />
-            </div>
-            <img
-              src="https://ghchart.rshah.org/7a0ba5/egrodo"
-              alt="My Github contributions this year"
-              title="My Github contributions this year"
-              className={imgLoaded ? '' : CSS.hide}
-              onLoad={() => this.setState({ imgLoaded: true })}
-            />
           </div>
-          <div className={CSS.postsArea}>
-            {edges.map(({ node }) => <ContentBlock type="project" node={node} key={node.title} />)}
-          </div>
-          <ContactSection />
         </div>
-      </section>
-    );
-  }
+        <div
+          className={CSS.bottomHeader}
+          style={{ display: 'none', visibility: 'hidden' }}
+          ref={fakeStickyRef}
+        >
+          <h2 className={CSS.postTitle}>Projects</h2>
+        </div>
+        <div className={CSS.bottomHeader} ref={stickyRef}>
+          <h2 className={CSS.postTitle}>Projects</h2>
+        </div>
+      </header>
+      <div className={CSS.content}>
+        <p className={CSS.indexBlurb}>
+          Below are a few personal projects I&apos;ve done that I&apos;m proud of, to see more check out my&nbsp;
+          <FancyLink to="https://github.com/egrodo" newTab>
+            Github page
+          </FancyLink>
+          !
+        </p>
+        <div className={CSS.GHChart}>
+          <div className={`${CSS.imgPlaceholder} ${imgLoaded ? CSS.hide : ''}`}>
+            <Spinner />
+          </div>
+          <img
+            src="https://ghchart.rshah.org/7a0ba5/egrodo"
+            alt="My Github contributions this year"
+            title="My Github contributions this year"
+            className={imgLoaded ? '' : CSS.hide}
+            onLoad={() => setLoaded(true)}
+          />
+        </div>
+        <div className={CSS.postsArea}>
+          {edges.map(({ node }) => <ContentBlock type="project" node={node} key={node.title} />)}
+        </div>
+        <ContactSection />
+      </div>
+    </section>
+  );
 }
 
 projects.propTypes = {
